requirements.txt
dice~=1.0.2
discord.py~=0.9.1
docopt~=0.6.2
pyparsing~=2.0.7
requests~=2.9.1
ws4py~=0.3.4
